==> notebooks/01_data_validation.ipynb <==
# Data validation
This is a mostly manual section, where we check data validity and inconsistencies. Also, it is important for the data to be resilient to bugs.

import os
import pickle as pkl
import re
import sys
import warnings

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns

# local packages
import customtools.data_wrangle as dw

# %matplotlib inline
warnings.filterwarnings('ignore')

help(dw)

# define functions

# get numerical columns
def get_numerical_mask(df):
    type_mask = []
    for i in df.dtypes:
        if str(i).startswith('float') or str(i).startswith('int'): # or str(i).startswith('bool')
            type_mask.append(True)
        else: type_mask.append(False)
    num_cols = list(np.array(df.columns)[type_mask])
    other_cols = list(np.array(df.columns)[[not elem for elem in type_mask]])
    
    return num_cols, other_cols

def duplicate_cleanup(df, col_id):
    original_size = df.shape[0]
    col_subset = df.columns.to_list()
    n_duplicates = df[(df.duplicated(col_subset))].shape[0]
    df.drop_duplicates(col_subset, inplace=True)
    print('number of removed duplicates:', n_duplicates)
    
    true_duplicates = df[(df.duplicated(col_id))].shape[0]
    if true_duplicates != 0:
        raise ValueError('There are still duplicates to verify')
        
    return df

temp = pd.DataFrame([[1, 2], [1, 2], [3, 4]], columns=['a', 'b'])

duplicate_cleanup(temp, 'a').values

# Define paths and capture data

inputs = os.path.join('..', 'data', '01_raw')
outputs = os.path.join('..', 'data', '02_intermediate')
reports = os.path.join('..', 'data', '06_reporting')

data = pd.read_csv(os.path.join(inputs, 'WA_Fn-UseC_-Telco-Customer-Churn.csv'))
# data_identity = pd.read_csv(os.path.join(inputs, 'train_identity.csv'))

print('shape of data:', data.shape[0])
data.head()

# standardize data

#### columns

data.rename({c: c.lower() for c in data.columns}, axis=1, inplace=True)
data.rename({'churn': 'y', 'customerid': 'id'}, axis=1, inplace=True)

#### strip values of categorical

for cat in data.columns:
    if data[cat].dtypes == 'object':
        data[cat] = data[cat].apply(lambda x: x.strip().lower())

# check duplicates for main data

data = duplicate_cleanup(data, 'id')

# set index

data.set_index('id', inplace=True)

# build new variables

# todo

# drop unused columns

# boolean features to float (0,1)

boolean_vars = ['partner', 'dependents', 'phoneservice', 'onlinesecurity', 'onlinebackup', 'deviceprotection', 
            'techsupport', 'streamingmovies', 'paperlessbilling', 'y']
for col in boolean_vars:
    data[col] = data[col].map( {'no': 0, 'yes': 1} ).astype(float)
    
data['y'] = data['y'].astype(int)
data['gender'] = data['gender'].map( {'female': 0, 'male': 1} ).astype(float)
data.rename({'gender': 'gender_male'}, axis=1, inplace=True)

# fix non-categorical types

data.loc[data['totalcharges'] == '', ['totalcharges']] = np.nan
data['totalcharges'] = data['totalcharges'].astype('float')

# evaluate categorical features

categories_dict = {}
for cat in data.columns:
    if data[cat].dtypes == 'object':
        categories_dict[cat] = list(data[cat].unique())
        n_categories = data[cat].unique().size
        print("number of categories for variable '{}': {} ".format(cat, n_categories))

# report data types

numerical_cols, other_cols = get_numerical_mask(data)
# numerical_cols.remove('y')
print(numerical_cols)

num_df = pd.DataFrame(numerical_cols)
other_df = pd.DataFrame(other_cols)

#### concatenate

report_df = pd.concat([num_df,other_df], ignore_index=True, axis=1)
report_df.columns = ['numerical_cols', 'non_numerical_cols']

# Checking  for possible anomalies in the database
Here we check for some possible anomalies. If there was one, we should look for solutions such as: contact the team responsible for data engineering or anyone who has a good understanding of the provided data, search for flaws in the data pipeline, or at least document for those anomalies.

Check if there is any oddity in data.

data.loc[:, ].describe()

# save data

data.head()

data.to_csv(os.path.join(outputs, 'data.csv'))
