==> notebooks/02_split.ipynb <==
# Split

import os
import sys
import warnings

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns
from sklearn.model_selection import train_test_split

# local packages
import customtools.data_wrangle as dw

warnings.filterwarnings('ignore')

# Define paths and capture data

inputs = os.path.join('..', 'data', '02_intermediate')
outputs = os.path.join('..', 'data', '02_intermediate')
reports = os.path.join('..', 'data', '06_reporting')

data = pd.read_csv(os.path.join(inputs, 'data.csv'), index_col='id')

print('Dataset dimensions:', data.shape)
data.head()

print('Dataset dimensions:', data_test.shape)
data_test.head()

# data sampling
if sampling_data == True: faster to run project, but will work on a data sample.

During sampling, we need to check if the distributions are similar to the original data by setting a significance level and using Kolmogorov-Smirnoff method. See function defined in the beginning of the notebook.

#from scipy.stats import ks_2samp

sampling_data = False

if sampling_data:
    data = dw.get_sample(data, significance=0.05, sample_size=5000, iterations=10)
data.shape

# final description

data.tail()

data.describe()

# Split data
test_size could depend on data size. For instance, for 1 million entries, it would work fine to establish test_size=0.1.

X = data.drop('y', axis=1)
print('dimensions of X:', X.shape)

y = data.loc[:, 'y']
y = y.astype('float')
print('dimensions of y:', y.shape)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, shuffle = False, random_state=42)
print('dimensions of X_train:', X_train.shape)
print('dimensions of y_train:', y_train.shape)
print('dimensions of X_test:', X_test.shape)
print('dimensions of y_test:', y_test.shape)

# save train and test sets

y_train = pd.DataFrame(y_train, columns=['y'])
y_test = pd.DataFrame(y_test, columns=['y'])

X_train.to_csv(os.path.join(outputs, 'X_train.csv'))
X_test.to_csv(os.path.join(outputs, 'X_test.csv'))
y_train.to_csv(os.path.join(outputs, 'y_train.csv'))
y_test.to_csv(os.path.join(outputs, 'y_test.csv'))
