==> notebooks/05_feature_engineering_sandbox.ipynb <==
# Feature Engineering

import os
import re
import statsmodels.api as sm
import sys
import warnings

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns
from sklearn.preprocessing import PolynomialFeatures

%matplotlib inline
warnings.filterwarnings('ignore')

sys.path.append(os.path.join('..', 'src'))

import importlib
import utils, params
importlib.reload(utils)
from utils import build_data_dict

importlib.reload(params)
from params import ProjectParameters

# define functions 

def build_polynomials(df, cols, method = 'simple_square', degrees=2):  
    print('number of columns before building polynomials:', df.shape[1])
    if method == 'simple_square':
        poly = df[cols]**2
        poly.columns = [c+'_power2' for c in df[cols].columns]
        df = pd.concat([df, poly], axis=1)
    elif method == 'k_degree':
        poly = PolynomialFeatures(2, include_bias = False)
        transformed = poly.fit_transform(df[cols])
        expanded_cols = poly.get_feature_names(df.columns)
        df = pd.DataFrame(transformed, columns = expanded_cols)
    print('number of columns after building polynomials:', df.shape[1])
    
    return df

def get_distribution(column, save_file):
    #Check the new distribution 
    sns.distplot(column, fit=norm);

    # Get the fitted parameters used by the function
    (mu, sigma) = norm.fit(column)

    # plot the distribution
    plt.legend(['Normal dist. ($\mu=$ {:.2f} and $\sigma=$ {:.2f} )'.format(mu, sigma)],
                loc='best')
    plt.ylabel('Frequency')
    plt.title('estimate distribution')
    plt.savefig(os.path.join(reports,save_file+'.jpg'), bbox_inches = "tight")
    plt.close()
    
    fig = plt.figure()
    res = stats.probplot(column, plot=plt)
    plt.savefig(os.path.join(reports,save_file+'_regline.jpg'), bbox_inches = "tight")
    plt.close()

### functions to deal with skewness
* bc_lambda is the box cox transformation on the variables. bc_lambda = 0 is equivalent to log transformation.

def get_skewness(df):
    numeric_features = df.dtypes[(df.dtypes != "object") & (df.dtypes != "bool")].index.to_list()
    skewed_features = df[numeric_features].apply(lambda x: skew(x.dropna())).sort_values(ascending=False)
    skewness = pd.DataFrame({'skew' :skewed_features})
    
    return skewness

def treat_skewness(df, df_name, bc_lambda = 0):
    skewness = get_skewness(df)
    print(skewness.head(10))

    skewness = skewness[skewness['skew'] > 0.75]
    print("There are {} skewed features prone to Box Cox transform on {} data".format(skewness.shape[0], df_name))

    skewed_features = skewness.index

    print(df_name)
    for feature in skewed_features:
        df[feature] = boxcox1p(df[feature], bc_lambda)
        
    return df

# Define paths

inputs = os.path.join('..', 'data', '02_intermediate')
outputs = os.path.join('..', 'data', '03_processed')
reports = os.path.join('..', 'data', '06_reporting')

# Data capture

data_list = ['X_train', 'X_test', 'X_train_oh', 'X_test_oh', 'y_train', 'y_test']

dfs_dict = build_data_dict(inputs, data_list)
for df in data_list:
    print(dfs_dict[df].shape)

for k,value in dfs_dict.items():
    print('shape of', k, 'is:', value.shape)

dfs_dict['X_train'].head()

dfs_dict['X_train'].describe()

dfs_dict['X_test'].describe()

# define continuous variables

numerical_cols = ProjectParameters().numerical_cols

# transformations

### polynomial features
* We might build simple polynomials with the square of continuous data
* another approach would be by using all possibilities from a degree of k. sklearn ref application: https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.PolynomialFeatures.html

test = build_polynomials(dfs_dict['X_train'], numerical_cols, method = 'simple_square', degrees=2)

for data in ['X_train', 'X_test', 'X_train_oh', 'X_test_oh']:
    dfs_dict[data] = build_polynomials(dfs_dict[data], numerical_cols, method = 'simple_square', degrees=2)

### symmetry

from scipy import stats
from scipy.stats import norm, skew
from scipy.special import boxcox1p, logit

# change skewness

Now let's see try to adjust skewness in the distribution.

transformed_dict = {}

### independent variables

boxcox_lambda = 0
for df_name in ['X_train', 'X_test']:
    transformed_dict[df_name] = treat_skewness(dfs_dict[df_name], df_name, boxcox_lambda)

### see how skewness changed

print('skewness after transformation:\r\n')
for df_name in ['X_train', 'X_test']:
    print(get_skewness(transformed_dict[df_name]).head(10))

# save data 

for df in ['X_train', 'X_test']:
    transformed_dict[df].to_csv(os.path.join(outputs, df+'.csv'))
    dfs_dict[df].to_csv(os.path.join(outputs, df+'_skewed.csv'))
for df in ['y_train', 'y_test']:
    dfs_dict[df].to_csv(os.path.join(outputs, df+'.csv'))
